+++ CHANGELOG.md
@@ -7,6 +7,11 @@ the resources to issue any form
 of guarantee that this log will cover all changes, and I will not attempt to go back
 very far in history.
 
+## Version [0.10.9] - 2023-03-01
+
+Silenced float('nan') and numpy.float64('nan) differences when validating values
+input against tsdf.values.
+
 ## Version [0.10.8] - 2023-02-28
 
 Using [Pydantic](https://docs.pydantic.dev/) for data validation using Python type